+++ spec/dummy/config/application.rb
@@ -15,6 +15,8 @@ module Dummy
     # Application configuration should go into files in config/initializers
     # -- all .rb files in that directory are automatically loaded.
 
+    config.middleware.use 'Apartment::Elevators::Subdomain'
+
     # Custom directories with classes and modules you want to be autoloadable.
     # config.autoload_paths += %W(#{config.root}/extras)