+++ spec/spec_helper.rb
@@ -5,6 +5,9 @@ ENV["RAILS_ENV"] = "test"
 
 require File.expand_path("../dummy/config/environment.rb",  __FILE__)
 require "rspec/rails"
+require 'capybara/rspec'
+require 'capybara/rails'
+
 
 ActionMailer::Base.delivery_method = :test
 ActionMailer::Base.perform_deliveries = true
@@ -17,5 +20,7 @@ Dir["#{File.dirname(__FILE__)}/support/**/*.rb"].each { |f| require f }
 
 
 RSpec.configure do |config|
+  
+  config.include RSpec::Integration::CapybaraSessions, :type => :request
   
 end