+++ spec/support/apartment_helpers.rb
@@ -9,5 +9,16 @@ module Apartment
     def self.drop_schema(schema)
       ActiveRecord::Base.connection.execute("DROP SCHEMA IF EXISTS #{schema} CASCADE")
     end
+    
+    def self.create_schema(schema)
+      ActiveRecord::Base.connection.execute("CREATE SCHEMA #{schema}")
+    end
+    
+    def self.in_database(db)
+      Apartment::Database.switch db
+      yield if block_given?
+      Apartment::Database.reset
+    end
+    
   end
 end